Rest of lelu-related model id:s from long to Integer

+++ src/main/java/fi/vaylavirasto/sillari/model/AxleChartModel.java
@@ -5,7 +5,7 @@ import java.util.List;
 
 @Data
 public class AxleChartModel {
-    private long id;
-    private long permitId;
+    private Integer id;
+    private Integer permitId;
     private List<AxleModel> axles;
 }
