+++ src/main/java/fi/vaylavirasto/sillari/model/AxleModel.java
@@ -4,8 +4,8 @@ import lombok.Data;
 
 @Data
 public class AxleModel {
-    private long id;
-    private long axleChartId;
+    private Integer id;
+    private Integer axleChartId;
     private Integer axleNumber;
     private Double weight;
     private Double distanceToNext;
