+++ src/main/java/fi/vaylavirasto/sillari/model/TransportDimensionsModel.java
@@ -2,12 +2,10 @@ package fi.vaylavirasto.sillari.model;
 
 import lombok.Data;
 
-import java.util.List;
-
 @Data
 public class TransportDimensionsModel {
-    private long id;
-    private long permitId;
+    private Integer id;
+    private Integer permitId;
     private Double height;
     private Double width;
     private Double length;
